pyproject.toml
[tool.poetry]
name = "assignment"
version = "0.1.0"
description = ""
authors = ["donghwan <ldh793@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "3.2.12"
mysqlclient = "^2.1.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
